requirements.txt
dnspython==2.6.1
pyfiglet==0.8.post0
